+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog - Agent Framework Toolkit
 
-## Unreleased
-- Added `OpenAIEmbeddingModels` collection of Open AI Embedding Models
+## Version 1.0.0-preview.251230.1
+- Added `OpenAIEmbeddingModels` const-collection of Open AI Embedding Models
 - Exposed `Connection` as a Property on all Factories
 - NuGet: Changed Project URL to be the [Wiki](https://github.com/rwjdk/AgentFrameworkToolkit/wiki)
 - Upgraded 'Google_GenerativeAI.Microsoft' NuGet from 3.4.1 to 3.6.1